Menuler kısmı ayarlanması gerceklestirildi

+++ app/Http/Controllers/api/back/menuler/indexController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\api\back\menuler;
 
+use App\Helpers\myHelper;
 use App\Http\Controllers\Controller;
 use App\Models\EkipModel;
 use App\Models\MenuModel;
@@ -11,7 +12,31 @@ use Yajra\DataTables\DataTables;
 
 class indexController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            if (myHelper::yetkiKontrol('menuler',"aktiflik")===false){
+                return redirect()->route('back.home.index')->with(array(
+                    "type" => "error",
+                    "title" => "Hata",
+                    "text" => "Yetkiniz Yok"
+                ));
+            }
+            return $next($request);
+        });
+    }
+
     public function index(){
+
+        if (myHelper::yetkiKontrol('menuler',"listeleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $query = MenuModel::query();
         $data = DataTables::of($query)
             ->addIndexColumn()
@@ -46,6 +71,15 @@ class indexController extends Controller
 
     // EKLEME KISMI
     public function store(Request $request){
+
+        if (myHelper::yetkiKontrol('menuler',"ekleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = $request->except("_token");
         $sonuc = MenuModel::create($data);
 
@@ -67,16 +101,43 @@ class indexController extends Controller
 
     // GUNCELLEME KISMI
     public function edit(MenuModel $item){
+
+        if (myHelper::yetkiKontrol('menuler',"guncelleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         return response()->json($item);
     }
 
     // GORUNTULEME KISMI
     public function show(MenuModel $item){
+
+        if (myHelper::yetkiKontrol('menuler',"goruntuleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         return response()->json($item);
     }
 
     // GUNCELLEME ISLEMI
     public function update(Request $request,MenuModel $item){
+
+        if (myHelper::yetkiKontrol('menuler',"guncelleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = $request->except("_token");
 
         $sonuc = $item->update($data);
@@ -99,6 +160,15 @@ class indexController extends Controller
 
     // SILME KISMI
     public function delete(MenuModel $item){
+
+        if (myHelper::yetkiKontrol('menuler',"silme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $sonuc = $item->delete();
 
         if ($sonuc) {
@@ -120,6 +190,15 @@ class indexController extends Controller
     // SIRALAMA KISMI AYARLANAMSI
     public function rankSetter(Request $request)
     {
+
+        if (myHelper::yetkiKontrol('menuler',"guncelleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         parse_str($request->post('data'), $sirala);
         $sirala = $sirala['item'];
 
@@ -133,6 +212,14 @@ class indexController extends Controller
     // AKTIF ĞPASFO KISMI AYARLANSIM
     public function isActiveSetter(Request $request, MenuModel $item)
     {
+        if (myHelper::yetkiKontrol('menuler',"guncelleme")===false){
+            return redirect()->route('back.home.index')->with(array(
+                "type" => "error",
+                "title" => "Hata",
+                "text" => "Yetkiniz Yok"
+            ));
+        }
+
         $data = ($request->data == "true") ? 1 : 0;
         $item->update(array(
             "menu_durum" => $data
